requirements.txt
flask-mongoengine==1.0.0
mongoengine==0.22.1
responses==0.13.3
mongomock==3.22.1
pymongo==3.11.3
Flask==1.1.2
GitPython
jinja2==2.11.3
typer
pydantic
itsdangerous==1.1.0
Werkzeug==1.0.1
click==7.1.2
urllib3<2.0.0
MarkupSafe==2.0.1
